requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.3
django-cors-headers==4.9.0
django-redis==5.4.0
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
drf-spectacular==0.29.0
gevent==25.5.1
greenlet==3.3.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
psycopg2==2.9.10
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.3
redis==6.2.0
referencing==0.37.0
requests==2.32.4
rich==14.2.0
rpds-py==0.30.0
six==1.17.0
SQLAlchemy==2.0.46
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
whitenoise==6.9.0
zope.event==6.1
zope.interface==8.1.1
